requirements.txt
fastapi
uvicorn
openai
opentelemetry-sdk
opentelemetry-exporter-otlp